Fixed troubles with multiple reference frames
Allocation of components done just once with all the components even all the
multiples to avoid reallocations.
In reference frames all the 3 component vectors and 3x3 matrices declared
and not allocated.
Changed the order in which the components are written in the output file:
all the multiple components are inserted where are created, and not at the
end of all the components.