Sources: fallback to non projection
When computing sources using plane-projected points on severely
warped elements (and small elements besides) the code failed, since
the original side is much longer than the projected one.

Now if this phenomenon happens the computation carries on with the original
points and not the projected ones to compute the sources. This generates
a quite big warning, but allows to carry on computations on bad quality meshes.